chore: portalicious: simplify change-password
Making some changes based on consistency and discussions had with Tal that weren't shared.

Namely:
- "user-settings" is now "change-password" everywhere
- change password submit button is never disabled. This is an a11y request from Tal (buttons should never be disabled - instead they should show errors when pressed).
- extra variables `changePasswordError` has been removed
- the variable `changePasswordFormSubmitted` is now being used the same way as in the login form - ie. it only signals when to show errors because the user clicked on the submit button in a disabled state
- template always shows form instead of changing based on state
- use toast for success message
- using form-validators instead of field-validators to simplify the validators themselves.
