+++ interfaces/Portalicious/src/app/pages/change-password/change-password.component.html
@@ -1,0 +1,121 @@
+<app-page-layout>
+  <div class="mb-6 flex items-end pt-6">
+    <h1
+      i18n="@@page-title-change-password"
+      class="min-w-48 text-3xl"
+    >
+      Change password
+    </h1>
+  </div>
+  <div
+    class="mt-4 flex w-full shrink-0 flex-col justify-center gap-6 bg-white p-4"
+  >
+    <form
+      class="max-w-xl"
+      [formGroup]="changePasswordForm"
+      (ngSubmit)="onChangePassword()"
+    >
+      <div class="flex flex-col gap-2">
+        <!-- eslint-disable-next-line @angular-eslint/template/label-has-associated-control -->
+        <label>
+          <ng-container i18n>Current Password</ng-container>
+          <p-password
+            formControlName="currentPassword"
+            [toggleMask]="true"
+            [feedback]="false"
+            autocomplete="current-password"
+            styleClass="w-full [&_input]:w-full mt-2"
+          />
+        </label>
+        <app-form-error
+          [visible]="
+            (changePasswordFormSubmitted() ||
+              changePasswordForm.controls.currentPassword.dirty) &&
+            changePasswordForm.controls.currentPassword.errors?.required
+          "
+          error="This field is required."
+          i18n-error="@@generic-required-field"
+        />
+        <!-- eslint-disable-next-line @angular-eslint/template/label-has-associated-control -->
+        <label>
+          <ng-container i18n>New Password</ng-container>
+          <p-password
+            formControlName="newPassword"
+            [toggleMask]="true"
+            [feedback]="false"
+            autocomplete="new-password"
+            styleClass="w-full [&_input]:w-full mt-2"
+          />
+        </label>
+        @if (
+          changePasswordFormSubmitted() ||
+          changePasswordForm.controls.newPassword.dirty
+        ) {
+          @if (changePasswordForm.controls.newPassword.errors?.required) {
+            <app-form-error
+              error="This field is required."
+              i18n-error="@@generic-required-field"
+            />
+          } @else if (
+            changePasswordForm.controls.newPassword.errors?.minlength
+          ) {
+            <app-form-error
+              error="The new password must be at least 8 characters long."
+              i18n-error
+            />
+          } @else {
+            <app-form-error
+              [visible]="
+                changePasswordForm.hasError('newPasswordIsNotDifferent')
+              "
+              error="The new password must be different from the current password."
+              i18n-error
+            />
+          }
+        }
+        <!-- eslint-disable-next-line @angular-eslint/template/label-has-associated-control -->
+        <label>
+          <ng-container i18n>Confirm Password</ng-container>
+          <p-password
+            formControlName="confirmPassword"
+            [toggleMask]="true"
+            [feedback]="false"
+            styleClass="w-full [&_input]:w-full mt-2"
+          />
+        </label>
+        @if (
+          changePasswordFormSubmitted() ||
+          changePasswordForm.controls.confirmPassword.dirty
+        ) {
+          @if (changePasswordForm.controls.confirmPassword.errors?.required) {
+            <app-form-error
+              error="This field is required."
+              i18n-error="@@generic-required-field"
+            />
+          } @else {
+            <app-form-error
+              [visible]="
+                changePasswordForm.hasError('confirmPasswordDoesNotMatch')
+              "
+              error="The confirm password must be equal to the new password."
+              i18n-error
+            />
+          }
+        }
+        <p-button
+          label="Change password"
+          i18n-label
+          type="submit"
+          rounded
+          class="mt-8"
+          [loading]="changePasswordMutation.isPending()"
+        />
+      </div>
+
+      <app-form-error
+        [visible]="changePasswordMutation.isError()"
+        [error]="changePasswordMutation.failureReason()?.message"
+      />
+    </form>
+  </div>
+</app-page-layout>
